+++ eo_man/icons/image_gallary.py
@@ -113,3 +113,7 @@ class ImageGallery():
     def get_forward_mail(cls, size:tuple[int:int]=(32,32)) -> ImageTk.PhotoImage:
         return ImageGallery.get_image("mail-forward.png", size)
     
+    @classmethod
+    def get_pct14_icon(cls, size:tuple[int:int]=(32,32)) -> ImageTk.PhotoImage:
+        return ImageGallery.get_image("pct14.png", size)
+